refactor(android): merge PurchaselyV6Bridge into PurchaselyModule — single module
Removed the standalone `PurchaselyV6Bridge` object (PurchaselyV6Module.kt) and
inlined its logic directly into `PurchaselyModule` so the Android side exposes a
single native module.

- The 8 v6 @ReactMethod entry points (v6Preload/v6Display/v6Close/v6Back/
  v6RegisterInterceptor/v6UnregisterInterceptor/v6CompleteInterceptor/
  v6ApplyStartOptions) now hold the implementation directly instead of
  delegating to PurchaselyV6Bridge — the JS contract is unchanged.
- v6 helpers (buildV6Prepared, wireV6Callbacks, toV6Map/toV6Payload/toV6String/
  toV6Ordinal) are now private members; they reuse the module's existing
  `sendEvent` and companion `transformPlanToMap` (no duplication).
- Event-name constants, the interceptor timeout, and per-request state
  (activeV6Requests, pendingV6Interceptors) live in the companion object to
  preserve the process-global semantics the former object singleton had.
- Deleted packages/.../reactnativepurchasely/v6/PurchaselyV6Module.kt and the
  now-empty v6/ package directory. No remaining references to PurchaselyV6Bridge.

Verified: brace balance 276/276, no orphaned references, imports complete;
yarn test (133 ✓) / typecheck ✓ / lint ✓. Kotlin not compilable in this
environment (native 6.0.0 SDK unpublished) — verified structurally.
